requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
mpmath==1.2.1
password-validator==1.0
phonenumbers==8.12.42
Pillow==9.0.1
psycopg2==2.9.3
qrcode==7.3.1
requests==2.27.1
sqlparse==0.4.2
sympy==1.9
tencentcloud-sdk-python==3.0.578
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
